requirements.txt
black==25.11.0
cfgv==3.5.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
duckdb==1.4.2
filelock==3.20.0
fonttools==4.60.1
identify==2.6.15
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.7
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.5.0
pyarrow==22.0.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
ruff==0.14.7
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
virtualenv==20.35.4
wheel==0.45.1
